[TASK] Add support for HTTP proxies
If TYPO3 is configured to use a proxy server for outgoing traffic,
the API requests to cleverreach will fail.

With this patch, the proxy server configured in TYPO3 install tool
will be used for all curl requests.

On a long run, it will be better to get rid of the custom implemenation
for HTTP requests and use Guzzle (bundled in TYPO3 Core since V8)
instead.